+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
   "crates/cairoc",
   "crates/flo",
   "crates/rust-test-input",
+  "crates/test-utils",
 ]
 
 # Cairo is excluded because it is imported as git submodule and it has its own Cargo workspace.
@@ -48,6 +49,7 @@ hieratika-compiler = { path = "crates/compiler" }
 hieratika-driver = { path = "crates/driver" }
 hieratika-errors = { path = "crates/error" }
 hieratika-flo = { path = "crates/flo" }
+hieratika-test-utils = { path = "crates/test-utils" }
 thiserror = "1.0.63"
 tracing = "0.1.40"
