+++ crates/compiler/tests/compilation_basic_add.rs
@@ -4,9 +4,7 @@
 use hieratika_flo::types::{BlockExit, Type};
 use itertools::Itertools;
 
-// Setting module as public to avoid `dead_code` warning for unused common test
-// functions.
-pub mod common;
+mod common;
 
 #[test]
 fn compiles_add() -> anyhow::Result<()> {
