+++ crates/test-utils/README.md
@@ -1,0 +1,3 @@
+# Hieratika Test Utils
+
+Handy functions to simplify testing of Hiratika.
