+++ crates/test-utils/input/square.cairo
