+++ crates/test-utils/input/square.ll
